Refactor(Lumina): Restructuring app
