+++ models/manager.models.ts
